Ajout des premiers capteurs

+++ NeuralNetwork/Model/Honey.cs
@@ -19,8 +19,8 @@ namespace NeuralNetwork.Model
         }
 
         #region Coordonnées
-        private int _x;
-        public int X
+        private double _x;
+        public double X
         {
             get { return this._x; }
             set
@@ -32,8 +32,8 @@ namespace NeuralNetwork.Model
                 }
             }
         }
-        private int _y;
-        public int Y
+        private double _y;
+        public double Y
         {
             get { return this._y; }
             set
